Fix windows cross compile with new netlink
